Network/Labs/Scripts/analyzer/requirements.txt
bootstrap
jQuery
Flask
Flask-WTF
pyx
requests
scapy-python3 
